refactor project structure,
update readme,
add .ycm_extra_conf for YCM autocompletion,
remove info.sh,
cast implicitly,
use unsigned where appropriate

+++ source/decode.c
@@ -10,8 +10,8 @@
 
 char decode_character(unsigned long encoded_char) {
   // try decoding with cipher 2^0 to 2^8:
-  for (int cipher = 1; cipher <= 256; cipher *= 2) {
-    int test_case = encoded_char / cipher;
+  for (unsigned int cipher = 1; cipher <= 256; cipher *= 2) {
+    unsigned long test_case = encoded_char / cipher;
 
     // expecting ASCII: 65(A) - 90(Z) + 32(space):
     if ((test_case >= 65 && test_case <= 90) || test_case == 32) {
@@ -23,8 +23,8 @@ char decode_character(unsigned long encoded_char) {
 
 void verbose_decode(unsigned long encoded_char, char *buffer) {
   // try decoding with cipher 2^0 to 2^8:
-  for (int cipher = 1; cipher <= 256; cipher *= 2) {
-    int test_case = encoded_char / cipher;
+  for (unsigned int cipher = 1; cipher <= 256; cipher *= 2) {
+    unsigned long test_case = encoded_char / cipher;
 
     // expecting ASCII: 65(A) - 90(Z) + 32(space):
     if ((test_case >= 65 && test_case <= 90) || test_case == 32) {
